Export hits_corrector to components
